+++ CHANGELOG.md
@@ -8,6 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Changed
+
+### Fixed
+
+
+## [1.0.0-alpha19] - 2022-10-07
+
+### Added
+
 - Enabled images and documents sharing in conversation screen.  
 
 ### Changed